+++ src/vec.c
@@ -2,8 +2,8 @@
 
 #include <SDL.h>
 
-float Vec2_Magnitude(Vec2 vec) {
-    float x2, y2, mag;
+f32 Vec2_Magnitude(Vec2 vec) {
+    f32 x2, y2, mag;
 
     x2 = vec.x * vec.x;
     y2 = vec.y * vec.y;
@@ -15,7 +15,7 @@ float Vec2_Magnitude(Vec2 vec) {
 
 Vec2 Vec2_Normalize(Vec2 vec) {
     Vec2  norm;
-    float mag;
+    f32 mag;
 
     norm = Vec2_Zero;
     mag  = Vec2_Magnitude(vec);
@@ -28,10 +28,10 @@ Vec2 Vec2_Normalize(Vec2 vec) {
     return norm;
 }
 
-Vec2 Vec2_Rotate(Vec2 vec, float deg) {
+Vec2 Vec2_Rotate(Vec2 vec, f32 deg) {
     Vec2  rot   = Vec2_Zero;
-    float a_sin = SDL_sinf(deg);
-    float a_cos = SDL_cosf(deg);
+    f32 a_sin = SDL_sinf(deg);
+    f32 a_cos = SDL_cosf(deg);
 
     rot.x = vec.x * a_cos - vec.y * a_sin;
     rot.y = vec.y * a_cos + vec.x * a_sin;
@@ -39,7 +39,7 @@ Vec2 Vec2_Rotate(Vec2 vec, float deg) {
     return rot;
 }
 
-int Vec2_Equal(Vec2 a, Vec2 b) {
+i8 Vec2_Equal(Vec2 a, Vec2 b) {
     if (a.x == b.x && a.y == b.y) {
         return 1;
     }
